% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.asgl}
\alias{predict.asgl}
\title{Make predictions from an 'asgl' object}
\usage{
\method{predict}{asgl}(object, newx, s = NULL, type = c("link",
  "response"), ...)
}
\arguments{
\item{object}{Fitted \code{asgl} object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to
be made.}

\item{s}{Value(s) of the penalty tuning parameter at which predictions are to
be made. Default is the entire sequence used to fit the regularization
path.}

\item{type}{Type of predicted required. For \code{"gaussian"} models,
\code{"link"} and \code{"response"} give the same result: the
predicted response. For \code{"binomial"} models, \code{"link"} gives
the linear predictor and \code{"response"} gives the predicted
probabilities.}

\item{...}{Ignored.}
}
\description{
Makes predictions from a fitted 'asgl' object.
}
